Change made by : Johannes Pretorius (dominion It) 20 Oct 2012
Added Backup of files when they are edited. Backups are created in the folder
data/_backup/pages

If a backup exist for a specific page (from previous backup), then a menu option is
added with mouse hover named 'Restore Backup'. when it is clicked the backup
will be restored and removed.

This way allows for only 1 backup to be kept per page at a time.