Update galleries with new files names if image file is renamed.
This will update all galleries with new file name for image if name changed.

this also tries to fix some bugs in elFinder
https://github.com/Studio-42/elFinder/issues/417
https://github.com/Studio-42/elFinder/pull/416

NOTE : that for issue 416
I found a better reason (more logical) and solution for this

make sure mime_magic.magicfile is set else this gives problems on windows set to mime_magic.magicfile = "$PHP_INSTALL_DIR\magic.mime"

But the problem here is, IF elFinder crashes in any way weirdly when doing something
it does not close the session and one can get session logs, this results in
the Open Directory haning scenario. We need to hook into the shutdown and add
code that IF a session is active, it closes it. hopefully that will fix it.