improve homepage setting in Page Manager
5.1.1 introduces the $config key 'homepath_auto', which brings back the
former default behavior of using the first main menu item as homepage -
unless explicitly set otherwise.
Once the hompage is set to be a certain custom page via the 'Homepage'
link in Page Manager, a new unlock/disable button will appear next to
it. Clicking it will unset the custom homepage and re-enable the auto
mode.