add FronteEndFramework section to Addon.ini
If present in a theme's Addon.ini, a 'framework' key will be added to
the corresponding  layout entry in $gpLayouts upon installation.
'framework' will be an array containing all key-pairs defined in the
Addon.ini section. This will make a reliable server-side framework
detection way easier, if not to say 'possible'.