add loader overlay while saving
Occasional page corruption is a hard-to-track but serious problem. It
can ruin edits since a page was last published. Adding the loader
overlay while saving seems to be the last option. Fortunately, this is
hardly noticeable with a fast internet connection. I wish I could (or
will) find a better way.