New event 'editor_area:shown'
Can I ask for a new JS event when editor is shown.
The 'editor_area:loaded' is not sufficient to catch the moment when, e.g., the gallery editor is changed to the pure CKeditor or the plugin section editor and vice versa. 
I need it for my CKE_Themes plugin.